Change parameter type

+++ src/appleseed/renderer/kernel/lighting/pathtracer.h
@@ -100,7 +100,7 @@ class PathTracer
         const ShadingContext&   shading_context,
         const ShadingPoint&     shading_point);
 
-    const ShadingPoint& get_path_vertex(const int i) const;
+    const ShadingPoint& get_path_vertex(const size_t i) const;
 
   private:
     PathVisitor&                m_path_visitor;
@@ -935,7 +935,7 @@ bool PathTracer<PathVisitor, VolumeVisitor, Adjoint>::march(
 }
 
 template<typename PathVisitor, typename VolumeVisitor, bool Adjoint>
-inline const ShadingPoint& PathTracer<PathVisitor, VolumeVisitor, Adjoint>::get_path_vertex(const int i) const
+inline const ShadingPoint& PathTracer<PathVisitor, VolumeVisitor, Adjoint>::get_path_vertex(const size_t i) const
 {
     return reinterpret_cast<const ShadingPoint*>(m_shading_point_arena.get_storage())[i];
 }
